implement comment docs parsing and templates to build the doc page on the fly.
The implementation is at the top of index.html and is a rough prototype.  I've converted
map, reduce, and reduceRight over to use it.  So now the description that was on the
index page is inline the function body in comments.  For now on the doc page
I just do a `document.write(InlineDocs(_.map))` to print out the identical html.

Waiting to get approval before converting all documentation over to this, as well as
generating the table of contents, version info, etc. dynamically.

Putting the docs inside the function body is not standard, but makes it very simple
to get at them with `fn.toString()`.

Beyond making maintenance of the documentation page much easier and perhaps more interactive,
the metadata extracted from the comments can be used for testing too.  One way is we can now
generate duplicate tests for all aliases of a function.